add: move check for turn in files in the interface
