+++ internal/tests/R00/ex04.go
@@ -67,9 +67,6 @@ func testCargoRun(exercise Exercise.Exercise) Exercise.Result {
 }
 
 func ex04Test(exercise *Exercise.Exercise) Exercise.Result {
-	if !testutils.TurnInFilesCheck(*exercise) {
-		return Exercise.Result{Passed: false, Output: "invalid files found in turn in directory"}
-	}
 	exercise.TurnInFiles = testutils.FullTurnInFilesPath(*exercise)
 
 	if result := testCargoRun(*exercise); !result.Passed {
